feat: se agrega pagina de registro de usuarios
